refactor(debt): extract runtime/toolchain.rs from runtime.rs god file (D-01 step 1)
Move repair_rootfs_hardlinks, force_symlink, and prepare_toolchain_wrappers
(plus their nested helpers) verbatim into a self-contained runtime/toolchain.rs
submodule. runtime.rs 2136 -> 1592 LOC. The cluster has no outgoing dependency
on the rest of runtime; re-exported via 'mod toolchain; use toolchain::{...}'.

Kept runtime.rs as the parent file (sibling submodule dir) rather than git mv
to runtime/mod.rs, to preserve blame on the remaining 1592 lines.

Verified under WSL Linux: cargo test --lib runtime 14/0; clippy --lib --tests
-D warnings clean. Two pre-existing repeat().take() warnings in the D-09 test
helpers fixed to resize() in passing (Boy-Scout).

Steps 2-3 (server.rs device-critical, extraction.rs) remain device-in-loop.
